crewai-service/requirements.txt
# CrewAI Service - Requirements (Simplified)

# Core
crewai==0.11.2
langchain-ollama

# Web Server
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
